test: add more unit tests
